requirements.txt
streamlit 
langchain_google_genai 
litellm 
pandas 
crewai 
matplotlib 
plotly
reportlab